python-services/rag_engine/requirements.txt
# RAG Engine Service Dependencies
langchain==0.0.340
langchain-openai==0.0.2
faiss-cpu==1.7.4
openai==1.3.7
tiktoken==0.5.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pydantic==2.5.0
numpy==1.24.3

# HuggingFace Dependencies
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
huggingface-hub==0.19.4
